requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
PyJWT==2.6.0
Email-Validator==1.1.3 